linnet::half_edge::hedgevec

Trait Accessors

pub trait Accessors<Index> {
    type Data;

    // Required methods
    fn orientation(&self, index: Index) -> Orientation;
    fn set_orientation(&mut self, index: Index, orientation: Orientation);
    fn pair(&self, index: Index) -> HedgePair;
    fn data(&self, index: Index) -> &Self::Data;
    fn data_mut(&mut self, index: Index) -> &mut Self::Data;
}

A trait providing a generic interface for accessing properties of edges (like orientation, data, and pairing information) using various index types.

This allows functions to be generic over how they access edge details, whether by Hedge, EdgeIndex, or HedgePair.

Required Associated Types

type Data

The type of the primary data associated with the edge.

Required Methods

fn orientation(&self, index: Index) -> Orientation

Returns the Orientation of the edge identified by index.

fn set_orientation(&mut self, index: Index, orientation: Orientation)

Sets the Orientation of the edge identified by index.

fn pair(&self, index: Index) -> HedgePair

Returns the HedgePair describing the pairing status of the edge/half-edge identified by index.

fn data(&self, index: Index) -> &Self::Data

Returns a reference to the custom data of the edge identified by index.

fn data_mut(&mut self, index: Index) -> &mut Self::Data

Returns a mutable reference to the custom data of the edge identified by index.