linnet::half_edge::involution

Struct Involution

pub struct Involution<E = EdgeIndex> { /* private fields */ }

Manages the topological relationships between half-edges in a graph.

An involution is a function f such that f(f(x)) = x. In this context, it maps each half-edge to its unique opposite half-edge. If a half-edge is unpaired (an “identity” or “external” edge), it maps to itself.

This struct stores a vector of InvolutiveMapping<E>, where each entry corresponds to a half-edge and describes its pairing and associated data.

Type Parameters

E: The type of data associated with each edge. Defaults to EdgeIndex if not specified, implying edges primarily store their own index as data, but typically this is instantiated with custom edge data for a HedgeGraph.

Implementations

impl<E> Involution<E>

pub fn display(&self) -> String
where E: Display,

pub fn inv(&self, hedge: Hedge) -> Hedge

The fundamental function of involutions is to pair edges. This function provides this pairing. It is a bijective map inv: H -> H, where H is the set of half-edges. The map is involutive, meaning that inv(inv(h)) = h. The fixed points of this function correspond to the unpaired hedges, interpreted as dangling or external edges.

pub fn is_empty(&self) -> bool

pub fn new() -> Self

pub fn add_identity( &mut self, data: E, orientation: impl Into<Orientation>, underlying: Flow, ) -> Hedge

add a new dangling edge to the involution returns the index of the new edge

pub fn connect_to_identity( &mut self, connect: Hedge, ) -> Result<Hedge, InvolutionError>

adds a new hedge that connects to an identity hedge

pub fn add_pair( &mut self, data: E, directed: impl Into<Orientation>, ) -> (Hedge, Hedge)

add a connected pair of hedges to the involution returns the pair of hedges: (source, sink)

pub fn find_from_data(&self, data: &E) -> Option<HedgePair>
where E: PartialEq,

pub fn last(&self) -> Option<Hedge>

pub fn connect_identities( &mut self, source_idx: Hedge, sink_idx: Hedge, merge_fn: impl FnOnce(Flow, EdgeData<E>, Flow, EdgeData<E>) -> (Flow, EdgeData<E>), ) -> Result<HedgePair, InvolutionError>
where E: Clone,

connect a pair of identity hedges, taking orientation and data from source_idx errors if the hedges are not identity

pub fn split_edge( &mut self, hedge: Hedge, data: EdgeData<E>, ) -> Result<(), InvolutionError>

Splits the edge that hedge is a part of into two dangling hedges, adding the data to the side given by hedge. The underlying orientation of the new edges is the same as the original edge, i.e. the source will now have Flow::Source and the sink will have Flow::Sink. The superficial orientation has to be given knowing this.

pub fn check_involutivity(&self) -> Result<(), String>

pub fn extract<S: SubSetLike, O>( &mut self, graph: &S, split_edge_fn: impl FnMut(EdgeData<&E>) -> EdgeData<O>, internal_data: impl FnMut(EdgeData<E>) -> EdgeData<O>, ) -> Involution<O>

Extracts the edges included in the subgraph into a separate involution. If the involution maps edges across the splitting boundary, i.e. if i in graph but inv(i) not in graph, then the edge involution is modified on each side, to turn this into a dangling edge. The data assigned to a fully owned edge is mapped by the internal_data closure, whilst those that get split, are assigned the data gotten from the split_edge_fn closure. The new dangling edges retain the old underlying orientation.

pub fn delete<S: SubSetLike>(&mut self, graph: &S)

pub fn hedge_pair(&self, hedge: Hedge) -> HedgePair

pub fn first_internal<S: SubSetLike>(&self, subgraph: &S) -> Option<Hedge>

pub fn n_internals<S: SubSetLike>(&self, subgraph: &S) -> usize

pub fn is_internal<S: SubSetLike>(&self, index: Hedge, subgraph: &S) -> bool

check if the edge is internal and totally included in the subgraph

pub fn orientation(&self, hedge: Hedge) -> Orientation

pub fn set_orientation(&mut self, hedge: Hedge, orientation: Orientation)

pub fn flow(&self, hedge: Hedge) -> Flow

pub fn iter_edge_data(&self) -> impl Iterator<Item = (Hedge, &EdgeData<E>)>

pub fn iter_edge_data_mut(&mut self) -> impl Iterator<Item = &mut EdgeData<E>>

pub fn iter(&self) -> InvolutionIter<'_, E>

pub fn iter_mut(&mut self) -> InvolutionIterMut<'_, E>

pub fn iter_idx(&self) -> impl Iterator<Item = Hedge>

pub fn as_ref(&self) -> Involution<&E>

pub fn map_data_ref<'a, G, E2>(&'a self, g: &G) -> Involution<E2>
where G: FnMut(&'a E) -> E2 + Clone,

pub fn map_data<G, E2>(self, g: &G) -> Involution<E2>
where G: FnMut(E) -> E2 + Clone,

pub fn map_data_option<F, G, E2>(self, g: &G) -> Option<Involution<E2>>
where G: FnMut(E) -> Option<E2> + Clone,

pub fn map_data_result<F, G, E2, O>(self, g: &G) -> Result<Involution<E2>, O>
where G: FnMut(E) -> Result<E2, O> + Clone,

pub fn map_full<E2>( self, g: impl FnMut(HedgePair, EdgeData<E>) -> EdgeData<E2>, ) -> Involution<E2>

pub fn is_sink(&self, hedge: Hedge) -> bool

pub fn is_source(&self, hedge: Hedge) -> bool

pub fn is_identity(&self, hedge: Hedge) -> bool

pub fn set_as_source(&mut self, hedge: Hedge) -> bool

If the data at hedge was a sink, turn it into a source. Otherwise do nothing.

pub fn set_as_sink(&mut self, hedge: Hedge)

If the data at hedge was a source, turn it into a sink. Otherwise do nothing.

pub fn edge_data(&self, index: Hedge) -> &EdgeData<E>

pub fn edge_data_mut(&mut self, index: Hedge) -> &mut EdgeData<E>

pub fn print<S: SubSetLike>( &self, subgraph: &S, h_label: &impl Fn(&E) -> Option<String>, ) -> String