linnet::half_edge::layout::simulatedanneale

Trait Energy

pub trait Energy<S> {
    // Required method
    fn energy(&self, prev: Option<(&S, f64)>, next: &S) -> f64;

    // Provided method
    fn on_accept(&self, _state: &mut S) { ... }
}

Required Methods

fn energy(&self, prev: Option<(&S, f64)>, next: &S) -> f64

Return the absolute energy of next. When a previous state/energy is provided, implementations may reuse or update it instead of recomputing from scratch.

Provided Methods

fn on_accept(&self, _state: &mut S)

Called once a proposal is accepted so implementations can clear or update caches.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".