linnet::half_edge

Module nodestore

Structs

BitVecNeighborIter

An iterator that yields the Hedge identifiers incident to a node, based on iterating over the set bits in a BitVec.

NodeStorageVec

An implementation of NodeStorage and NodeStorageOps that uses Vecs and BitVecs to store node information and their incident half-edges.

Traits

NodeStorage

The fundamental trait defining the interface for a node storage backend in a HedgeGraph.

NodeStorageOps

Defines a comprehensive set of operations for a node storage backend within a HedgeGraph.

Type Aliases

DefaultNodeStore