linnet::half_edge

Module subgraph

Re-exports

pub use contracted::ContractedSubGraph;

pub use cut::OrientedCut;

pub use cycle::Cycle;

pub use internal::InternalSubGraph;

pub use node::HedgeNode;

pub use empty::Empty;

pub use full::FullOrEmpty;

pub use subset::SuBitGraph;

Modules

contracted

cut

cycle

empty

full

internal

node

subset

Structs

SubSetIter

An iterator over Hedges included in a subgraph.

Traits

BaseSubgraph

Inclusion

A trait for checking inclusion or intersection relationships between a subgraph and another entity T.

ModifySubSet

Defines methods for modifying a subgraph by adding or removing elements.

PairwiseSubSetOps

Self-only combinators built on top of SubSetOps.

SubGraphLike

SubGraphOps

SubSetLike

The core trait defining the properties and behaviors of a subgraph representation.

SubSetOps

Defines binary set operations between a subgraph and another subset-like value.