linnet::half_edge::subgraph::cut

Type Alias CutGraph

pub type CutGraph<E, V, H, N> = HedgeGraph<PossiblyCutEdge<E>, V, H, N>;

Aliased Type

pub struct CutGraph<E, V, H, N> {
    pub node_store: N,
    /* private fields */
}

Fields
node_store: N

Storage for all nodes in the graph, including their data (V) and information about the half-edges incident to them. The specific implementation is determined by the S type parameter.

Implementations

impl<E, V, H, N: NodeStorageOps<NodeData = V>> CutGraph<E, V, H, N>

pub fn split(&mut self)

pub fn split_clone(&mut self)
where E: Clone,

pub fn split_copy(&mut self)
where E: Copy,

pub fn round_trip_split(&mut self)

pub fn round_trip_glue(&mut self)

pub fn cut(&self) -> OrientedCut

pub fn debug_cut_dot(&self) -> String

pub fn glue_back_strict(&mut self)

glues_back a cut graph. only matches edges with the same cut data, reversed cut flow, and compatible orientation and flow.

pub fn glue_back_lenient(&mut self)

glues_back a cut graph. only matches edges with the same cut data, reversed cut flow, and compatible orientation and without flow matching.