linnet::half_edge::subgraph

Trait BaseSubgraph

pub trait BaseSubgraph:
    SubGraphLike
    + ModifySubSet<HedgePair>
    + ModifySubSet<Hedge>
    + SubSetOps<Hedge> {
    // Required methods
    fn from_filter<E, V, H, N: NodeStorageOps<NodeData = V>, F: FnMut(&E) -> bool>(
        graph: &HedgeGraph<E, V, H, N>,
        filter: F,
    ) -> Self;
    fn from_hedge_iter<I: Iterator<Item = Hedge>>(iter: I, len: usize) -> Self;
}

Required Methods

fn from_filter<E, V, H, N: NodeStorageOps<NodeData = V>, F: FnMut(&E) -> bool>( graph: &HedgeGraph<E, V, H, N>, filter: F, ) -> Self

Selects all half-edges belonging to edges whose data satisfies filter.

fn from_hedge_iter<I: Iterator<Item = Hedge>>(iter: I, len: usize) -> Self

Creates a base subgraph containing all hedges yielded by the iterator. len specifies the total number of possible hedges in the graph context.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".