linnet::half_edge::subgraph

Trait Inclusion

pub trait Inclusion<T> {
    // Required methods
    fn includes(&self, other: &T) -> bool;
    fn intersects(&self, other: &T) -> bool;
}

A trait for checking inclusion or intersection relationships between a subgraph and another entity T.

This is used to determine if a subgraph contains certain elements (like specific hedges, ranges of hedges, or other subgraphs) or overlaps with them.

Type Parameters

T: The type of the entity to check for inclusion or intersection with.

Required Methods

fn includes(&self, other: &T) -> bool

Checks if this subgraph fully includes/contains the other entity.

fn intersects(&self, other: &T) -> bool

Checks if this subgraph has any common elements with the other entity.