linnet::half_edge::subgraph

Trait ModifySubSet

pub trait ModifySubSet<Index> {
    // Required methods
    fn add(&mut self, index: Index);
    fn sub(&mut self, index: Index);

    // Provided method
    fn union_with_iter(&mut self, other: impl Iterator<Item = Index>) { ... }
}

Defines methods for modifying a subgraph by adding or removing elements.

Type Parameters

Index: The type of element to add or remove (e.g., Hedge, HedgePair).

Required Methods

fn add(&mut self, index: Index)

Adds an element index to the subgraph.

fn sub(&mut self, index: Index)

Removes an element index from the subgraph.

Provided Methods

fn union_with_iter(&mut self, other: impl Iterator<Item = Index>)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.