linnet::tree

Module parent_pointer

Implements a tree structure where each node only stores a pointer to its parent.

Structs

PPNode

Represents a node within a ParentPointerStore.

ParentPointerStore

A forest data structure where each node only stores its data and a pointer to its parent.

Enums

ParentId

Identifies the parent of a node.