linnet::tree

Struct Forest

pub struct Forest<R, N> { /* private fields */ }

Represents a forest (a collection of disjoint trees).

R is the type of data associated with each root (tree). N is the storage implementation for the nodes, which must implement ForestNodeStore. Typically N will be one of ParentPointerStore<V>, ParentChildStore<V>, or ChildVecStore<V>, where V is the type of data associated with each node.

Implementations

impl<V, R> Forest<R, ParentPointerStore<V>>

pub fn change_to_root(&mut self, new_root: TreeNodeId) -> RootId

Changes the root of the tree containing new_root.

All parent pointers on the path from the new_root to the original root are reversed. The new_root becomes the root node associated with the original tree’s RootId. Updates the Forest’s root tracking accordingly.

Returns the RootId of the tree that was modified.

impl<R, N: ForestNodeStoreDown> Forest<R, N>

Methods available when the node store supports downward traversal.

pub fn iter_children( &self, start: TreeNodeId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns an iterator over the direct children of the node start.

impl<R, N: ForestNodeStoreRootLeaves> Forest<R, N>

Methods available when the node store supports iterating leaves of a specific root. (Note: This is automatically implemented if ForestNodeStoreDown is implemented).

pub fn iter_root_leaves( &self, root_id: RootId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns an iterator over all leaf nodes within the tree identified by root_id.

impl<R, N: ForestNodeStorePreorder> Forest<R, N>

Methods available when the node store supports pre-order traversal.

pub fn iter_preorder( &self, start: TreeNodeId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns a pre-order DFS iterator starting from the given node.

impl<R, N: ForestNodeStoreBfs> Forest<R, N>

Methods available when the node store supports BFS traversal.

pub fn iter_bfs( &self, start: TreeNodeId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns a BFS iterator starting at the given node.

impl<R, N: ForestNodeStoreAncestors> Forest<R, N>

Methods available when the node store supports ancestor traversal. (Note: This is automatically implemented for any ForestNodeStore).

pub fn iter_ancestors( &self, start_node: TreeNodeId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns an iterator that traverses upwards from start_node towards its root.

impl<R, N: ForestNodeStore> Forest<R, N>

pub fn root_node_for_root(&self, root_id: RootId) -> TreeNodeId

pub fn root_node_for_node(&self, node_id: NodeIndex) -> TreeNodeId

pub fn root_hedge_for_node(&self, node_id: NodeIndex) -> Hedge

impl<R, N> Forest<R, N>
where N: ForestNodeStore + Default + Into<ParentPointerStore<N::NodeData>> + From<ParentPointerStore<N::NodeData>>,

pub fn reroot_many( &mut self, new_roots: impl IntoIterator<Item = TreeNodeId>, ) -> usize

impl<R, N: ForestNodeStore> Forest<R, N>

General Forest methods available for any ForestNodeStore.

pub fn iter_roots(&self) -> impl Iterator<Item = (&R, &TreeNodeId)>

pub fn iter_root_ids(&self) -> impl Iterator<Item = RootId>

pub fn iter_nodes( &self, ) -> impl Iterator<Item = (TreeNodeId, Option<&N::NodeData>)>

pub fn iter_node_ids(&self) -> NodeIdIter

pub fn root(&self, nodeid: TreeNodeId) -> RootId

pub fn map_roots<R2>(self, transform: impl FnMut(R) -> R2) -> Forest<R2, N>

pub fn map_roots_ref<R2>( &self, transform: impl FnMut(&R) -> R2, ) -> Forest<R2, N>
where N: Clone,

pub fn map_nodes<F, T, U>(self, transform: F) -> Forest<R, N::Store<U>>
where N: ForestNodeStore<NodeData = T>, N::Store<U>: ForestNodeStore<NodeData = U>, F: FnMut(T) -> U,

pub fn map_nodes_ref<F, T, U>(&self, transform: F) -> Forest<R, N::Store<U>>
where N: ForestNodeStore<NodeData = T>, N::Store<U>: ForestNodeStore<NodeData = U>, F: FnMut(&T) -> U, R: Clone,

impl<R, N> Forest<R, N>

Methods for constructing/modifying the forest.

pub fn add_root<T>( &mut self, node_data: T, tree_data: R, ) -> (TreeNodeId, RootId)
where N: ForestNodeStore<NodeData = T>,

Adds a new root node to the forest, creating a new tree.

Takes the data for the new node (node_data) and the data for the new tree (tree_data). Returns the TreeNodeId of the new root node and the RootId of the new tree. Requires N to implement ForestNodeStore<NodeData = T>.

pub fn add_child<T>( &mut self, parent_id: TreeNodeId, node_data: T, ) -> TreeNodeId
where N: ForestNodeStore<NodeData = T>,

Adds a new child node to the specified parent node.

Takes the TreeNodeId of the parent and the data for the new child node (node_data). Returns the TreeNodeId of the newly added child node. Requires N to implement ForestNodeStore<NodeData = T>.

impl<R, N: Default> Forest<R, N>

pub fn new() -> Self

impl<U, P: ForestNodeStore> Forest<U, P>

pub fn root_node(&self, nodeid: TreeNodeId) -> TreeNodeId

pub fn validate_structure(&self) -> Result<(), ForestError>

pub fn debug_draw<FR, FN>(&self, format_root: FR, format_node: FN) -> String
where FR: FnMut(&U) -> String, FN: FnMut(&Option<P::NodeData>) -> String, P: ForestNodeStoreDown,

pub fn swap_roots(&mut self, a: RootId, b: RootId)

pub fn from_bitvec_partition<I: IntoIterator<Item = (U, SuBitGraph)>>( bitvec_part: I, ) -> Result<Self, ForestError>

impl<P, V> Forest<V, P>

pub fn cast<PP>(self) -> Forest<V, PP>
where PP: From<P>,