linnet::tree

Trait ForestNodeStore

pub trait ForestNodeStore:
    for<'a> Index<&'a TreeNodeId, Output = ParentId>
    + Index<TreeNodeId, Output = Option<Self::NodeData>>
    + IndexMut<TreeNodeId, Output = Option<Self::NodeData>>
    + Sized {
    type NodeData;
    type Store<T>: ForestNodeStore<NodeData = T> + From<ParentPointerStore<T>> + Into<ParentPointerStore<T>>;

    // Required methods
    fn to_store(self) -> Self::Store<Self::NodeData>;
    fn from_store(store: Self::Store<Self::NodeData>) -> Self;
    fn debug_draw(
        &self,
        node_display: impl FnMut(Option<&Self::NodeData>) -> Option<String>,
    ) -> String;
    fn validate(&self) -> Result<Vec<(RootId, TreeNodeId)>, ForestError>;
    fn n_nodes(&self) -> usize;
    fn swap(&mut self, a: TreeNodeId, b: TreeNodeId);
    fn set_root(&mut self, a: TreeNodeId, root: RootId);
    fn split_off(&mut self, at: TreeNodeId) -> Self;
    fn reparent(&mut self, parent: TreeNodeId, child: TreeNodeId);
    fn iter_nodes(
        &self,
    ) -> impl Iterator<Item = (TreeNodeId, Option<&Self::NodeData>)>;
    fn add_root(&mut self, data: Self::NodeData, root_id: RootId) -> TreeNodeId;
    fn add_dataless_root(&mut self, root_id: RootId) -> TreeNodeId;
    fn add_child(
        &mut self,
        data: Self::NodeData,
        parent: TreeNodeId,
    ) -> TreeNodeId;
    fn set_node_data(
        &mut self,
        data: Self::NodeData,
        node_id: TreeNodeId,
    ) -> Option<Self::NodeData>;
    fn add_dataless_child(&mut self, parent: TreeNodeId) -> TreeNodeId;
    fn map<F, U>(self, transform: F) -> Self::Store<U>
       where F: FnMut(Self::NodeData) -> U;
    fn map_ref<F, U>(&self, transform: F) -> Self::Store<U>
       where F: FnMut(&Self::NodeData) -> U;
    fn forgetful_map<U>(&self) -> Self::Store<U>;
    fn extend(&mut self, other: Self, shift_roots_by: RootId);

    // Provided methods
    fn parent(&self, child: TreeNodeId) -> ParentId { ... }
    fn panicing_validate(&self) { ... }
    fn from_pp(pp: ParentPointerStore<Self::NodeData>) -> Self { ... }
    fn make_root_of(&mut self, x: TreeNodeId, y: TreeNodeId) -> TreeNodeId { ... }
    fn root(&self, nodeid: TreeNodeId) -> RootId { ... }
    fn root_node(&self, nodeid: TreeNodeId) -> TreeNodeId { ... }
    fn iter_node_id(&self) -> NodeIdIter  { ... }

}

The core trait defining the interface for node storage within a Forest. Requires Index and IndexMut support for accessing node data and ParentId.

Required Associated Types

type NodeData

type Store<T>: ForestNodeStore<NodeData = T> + From<ParentPointerStore<T>> + Into<ParentPointerStore<T>>

Required Methods

fn to_store(self) -> Self::Store<Self::NodeData>

fn from_store(store: Self::Store<Self::NodeData>) -> Self

fn debug_draw( &self, node_display: impl FnMut(Option<&Self::NodeData>) -> Option<String>, ) -> String

fn validate(&self) -> Result<Vec<(RootId, TreeNodeId)>, ForestError>

fn n_nodes(&self) -> usize

fn swap(&mut self, a: TreeNodeId, b: TreeNodeId)

fn set_root(&mut self, a: TreeNodeId, root: RootId)

fn split_off(&mut self, at: TreeNodeId) -> Self

fn reparent(&mut self, parent: TreeNodeId, child: TreeNodeId)

fn iter_nodes( &self, ) -> impl Iterator<Item = (TreeNodeId, Option<&Self::NodeData>)>

fn add_root(&mut self, data: Self::NodeData, root_id: RootId) -> TreeNodeId

Adds a new root node.

fn add_dataless_root(&mut self, root_id: RootId) -> TreeNodeId

fn add_child(&mut self, data: Self::NodeData, parent: TreeNodeId) -> TreeNodeId

Adds a new child node as the last child of the parent.

fn set_node_data( &mut self, data: Self::NodeData, node_id: TreeNodeId, ) -> Option<Self::NodeData>

fn add_dataless_child(&mut self, parent: TreeNodeId) -> TreeNodeId

Adds a new child node as the last child of the parent.

fn map<F, U>(self, transform: F) -> Self::Store<U>
where F: FnMut(Self::NodeData) -> U,

fn map_ref<F, U>(&self, transform: F) -> Self::Store<U>
where F: FnMut(&Self::NodeData) -> U,

fn forgetful_map<U>(&self) -> Self::Store<U>

fn extend(&mut self, other: Self, shift_roots_by: RootId)

Provided Methods

fn parent(&self, child: TreeNodeId) -> ParentId

fn panicing_validate(&self)

fn from_pp(pp: ParentPointerStore<Self::NodeData>) -> Self

fn make_root_of(&mut self, x: TreeNodeId, y: TreeNodeId) -> TreeNodeId

Makes x’s root, the root of y’s root.

fn root(&self, nodeid: TreeNodeId) -> RootId

Finds the RootId by traversing upwards. Default implementation provided.

fn root_node(&self, nodeid: TreeNodeId) -> TreeNodeId

fn iter_node_id(&self) -> NodeIdIter

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.