linnet::tree

Trait ForestNodeStoreRootLeaves

pub trait ForestNodeStoreRootLeaves: ForestNodeStoreDown {
    // Required method
    fn iter_root_leaves(
        &self,
        root_id: RootId,
    ) -> impl Iterator<Item = TreeNodeId> + '_;
}

Trait for stores supporting iteration over leaves belonging to a specific root.

Required Methods

fn iter_root_leaves( &self, root_id: RootId, ) -> impl Iterator<Item = TreeNodeId> + '_

Returns an iterator over the leaf nodes belonging only to the tree with root_id.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".